requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
yfinance==0.2.40
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
gunicorn==21.2.0